stemming the tide of new 0.jpg.jpg user picture names
