+++ kifi-backend/modules/common/app/com/keepit/social/BasicUser.scala
@@ -3,6 +3,7 @@ package com.keepit.social
 import com.keepit.common.cache._
 import com.keepit.common.db._
 import com.keepit.common.logging.AccessLog
+import com.keepit.common.store.S3UserPictureConfig
 import com.keepit.model._
 import play.api.libs.functional.syntax._
 import play.api.libs.json._
@@ -60,7 +61,7 @@ object BasicUser {
       externalId = user.externalId,
       firstName = user.firstName,
       lastName = user.lastName,
-      pictureName = user.pictureName.map(_ + ".jpg").getOrElse("0.jpg"), // need support for default image
+      pictureName = user.pictureName.getOrElse(S3UserPictureConfig.defaultName) + ".jpg",
       username = user.username
     )
   }
